Working on `ExperimentSetup`
This is basically a combination of DaqPoints and and an output device.
The bsic interface is the function `movenext!` that setups the system
for the next measurement.

There are some tricky issues that I will have to give further thought.
The problem is that the parameters and axes names can be different and
it is possible to use transformations. I will try to keep it simple for now.