Simplified experimentsetup.jl
Now the paramters should be the same as the axes.
At most, the user can provide a map. If anything mor fancy
is needed, it should be done on the output device. Things such as
coordinates transformations should be handled by the output device or
a wrapper of the output device.