Implemented DaqSampling* and tests
Up 'til this point, the timing of data acquisition was simply
an initial acquisition time, a sampling rate and the number of samples.

Other types of timing might be necessary. Now we have AbstractDaqSampling
to model this. For now we have DaqSamplingRate that uses fixed sampling
rate. There is also DaqSamplingTimes that handles cases where each sample
is acquired at a specific point.

I also changed a bit the tests. The tests for each part the package
is now enclosed in a `let` block. This will avoid potential interference
between different tests: we only have to deal with one file at a time.