Improved a little bit the interface. Found a nasty bug in samplinghours and when converting from Rate -> Times
The problem is accuracy. DateTime has an accuracy of ms. If the sampling rate is high and is a broken value (fractional numbers, etc) when calculating the samplinghours there was an error and the number of samples could vary. Now this doesn't happen but there is a little loss of accuracy.